+++ cpu/src/buffer.cpp
@@ -6,6 +6,7 @@
 #include <stdlib.h>
 #include <sys/types.h>
 #include <sys/stat.h>
+#include <SDL2/SDL.h>
 
 #include "arrays.h"
 #include "type.h"
@@ -67,7 +68,7 @@ size_t text_size (FILE *file_text)
 	struct stat st = {};
     int fd = fileno (file_text); 
     fstat (fd, &st);
-    size_t size_text = st.st_size;
+    size_t size_text = (size_t) st.st_size;
 	return size_text;
 }
